backend/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.2
Flask-SocketIO>=5.3.2,<7
python-socketio==5.9.0
python-engineio==4.8.2
eventlet==0.33.3
gunicorn==21.2.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.4
Flask-Migrate==4.0.4
marshmallow==3.20.1
Werkzeug==2.3.8
Flask-Limiter==2.8.1
psycopg2-binary==2.9.7
alembic==1.11.1
numpy==1.26.4
scikit-image==0.20.0
Pillow==10.4.0
pycryptodome==3.18.0
